use uint for the FILTER_TABLE string len
it'd be retarded if a function name takes up more than what uint can
hold.

+++ config.def.h
@@ -51,7 +51,7 @@ static const enum output OUTPUT_DEFAULT = OUTPUT_ALL;
 /* convenient macro for populating the filter table */
 #define FILTER_TABLE_ENTRY(X) #X, sizeof (#X) - 1, X
 /* table of filter functions, used by filter_parse() for mapping --mag-filters */
-static const struct { const char *str; size_t len; FilterFunc f; } FILTER_TABLE[] = {
+static const struct { const char *str; uint len; FilterFunc f; } FILTER_TABLE[] = {
 	{ FILTER_TABLE_ENTRY(square) },
 	{ FILTER_TABLE_ENTRY(xhair)  },
 	{ FILTER_TABLE_ENTRY(grid)   },
